Computes the number of attacks incurred by selected airline in the selected year by damages level

This function computes the number of wildlife attacks incurred by airlines by damages level. This function is also being used in Shiny App.

Usage

damages_airline(input_operator, input_year)

Arguments

input_operator

Name of the airlines in the dataset, namely AMERICAN AIRLINES, UNITED AIRLINES, SOUTHWEST AIRLINES and DELTA AIR LINES

input_year

Year in the dataset, ranging from 1990 to 2018

Value

A dataframe.

Examples

damages_airline("AMERICAN AIRLINES", 2011)
#> # A tibble: 3 × 4
#> # Groups:   operator, damage, incident_year [3]
#>   operator          damage      incident_year     n
#>   <chr>             <fct>               <dbl> <int>
#> 1 AMERICAN AIRLINES Minor                2011     3
#> 2 AMERICAN AIRLINES Uncertain            2011     1
#> 3 AMERICAN AIRLINES Substantial          2011     2
damages_airline("UNITED AIRLINES", 2011)
#> # A tibble: 2 × 4
#> # Groups:   operator, damage, incident_year [2]
#>   operator        damage      incident_year     n
#>   <chr>           <fct>               <dbl> <int>
#> 1 UNITED AIRLINES Minor                2011     1
#> 2 UNITED AIRLINES Substantial          2011     1